PreyPredator-Game
 Prey-Predator simulation
In the next two laboratories, you will implement a prey-predator simulation application. The simulation will take place over several cycles C in an environment represented as an NxM grid. Each cell in this grid can be occupied by the following: Fox (F),  Gopher (G), Plant (P) or Empty (E).

When performing the simulation, you need to analyze the neighborhood of each cell; this neighborhood is defined as a 3 × 3 square centered in the current cell (in the right figure, the neighbors of the current cell, depicted in blue, are marked in red).  If you are dealing with a cell that is located on the margin of the grid, you need to ignore the neighbor cells that are outside the grid area ( x < 0 || x >= M, y < 0, y > N).
Survival rules
Empty
In the next iteration, an Empty cell will be replaced by:
a) Gopher, if more than one neighboring Gopher;
b) otherwise, Fox, if more than one neighboring Fox;
d) otherwise, Plant, if at least one neighboring Plant;
e) otherwise, Empty.


Plant
In the next iteration, the life form on a Plant cell will be transformed into:
a) Empty if there are twice as many Gophers as Plant cells in the neighborhood;
b) otherwise, Gopher if there are at least three Gophers in the neighborhood;
c) otherwise, Plant.


Gopher
In the next iteration, the life form on a Gopher cell will be transformed into:
a) Empty if the Gopher’s current age is 4 (the gopher dies of old age);
b) otherwise, Empty if there is no Plant in the neighborhood (the gopher starves to death);
c) otherwise, Fox if in the neighborhood there are at least as many Foxes as Gophers;
d) otherwise, Gopher(the gopher will survive).
Fox
In the next iteration, the life form on a Fox cell will be transformed into:
a) Empty if the Fox has an age greater or equal to 5 (the fox dies of old age);
b) otherwise, Empty, if there isn’t any Gopher in the neighborhood (the fox starves to death);
c) otherwise, Fox (the fox will live on).